requirements.txt
yt-dlp
pydub
PyYAML
openai
requests 